CodeQL fixes: real buffer overflow + format-string + overflow guards
Triaged the 29 CodeQL alerts on the repo. 19 are in third-party headers
(stb_image, stb_image_write, imgui internals) and not ours to fix. The
remaining 10 break down as:

#27 (error) - cpp/badly-bounded-write at img_io.cpp:1037. strncpy was
copying up to 63 bytes into IMG::src_filename which is char[16], with
a trailing src_filename[63] = '\0' that clobbered ~48 bytes past the
end of the struct. Real out-of-bounds write. The 63 was a leftover
constant from an older field layout. Fixed to use sizeof()-1, also
takes care of the two follow-on static-buffer-overflow warnings
(#28, #29) on the same site.

#26 (error) - cpp/wrong-number-format-arguments at imgui_overlay.cpp:8046.
SetTooltip is printf-style and the literal "25%" in the Onion-skin
tooltip was being read as a malformed format specifier. Escaped to "25%%".

#21-25 (warnings) - cpp/integer-multiplication-cast-to-long in five of
our own callsites: w*h or stride*h done as int/uint then assigned into
size_t. None can actually overflow for realistic sprite sizes (the
on-disk IMG geometry fields are 16-bit and the toolchain caps well
below the int max), but the operations aren't provably safe. Cast one
operand to size_t at each callsite so the multiplication itself lifts
to size_t. PoolAlloc takes size_t now to match - was unsigned int
which would have truncated at the API boundary.

+++ platform/img_format.h
@@ -196,7 +196,7 @@ static inline void FreePal(PAL *pal)
     free(pal);
 }
 
-static inline void *PoolAlloc(unsigned int n) { return calloc(1, n); }
+static inline void *PoolAlloc(size_t n) { return calloc(1, n); }
 #define PoolFree  free
 #define PoolDup(p, sz) memcpy(malloc(sz), (p), (sz))
